requirements.txt
absl-py==2.2.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
grpcio==1.71.0
joblib==1.5.0
kiwisolver==1.4.8
Markdown==3.8
matplotlib==3.10.3
packaging==25.0
protobuf==6.31.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torchvision==0.19.1
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
